split into api and mvc controller
