+++ src/AwesomeCMSCore/Modules/AwesomeCMSCore.Modules.Admin/Controllers/API/AccountController.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.Threading.Tasks;
+using AwesomeCMSCore.Modules.Admin.Extensions;
+using AwesomeCMSCore.Modules.Admin.Models.AccountViewModels;
+using AwesomeCMSCore.Modules.Entities.Entities;
+using AwesomeCMSCore.Modules.Helper.Email;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+
+namespace AwesomeCMSCore.Modules.Admin.Controllers.API
+{
+
+    [Authorize]
+    [Route("api/[controller]/[action]")]
+    public class AccountController : Controller
+    {
+        private readonly UserManager<User> _userManager;
+        private readonly SignInManager<User> _signInManager;
+        private readonly IEmailSender _emailSender;
+
+        public AccountController(
+            UserManager<User> userManager,
+            SignInManager<User> signInManager,
+            IEmailSender emailSender)
+        {
+            _userManager = userManager;
+            _signInManager = signInManager;
+            _emailSender = emailSender;
+        }
+
+
+        [HttpPost]
+        [AllowAnonymous]
+        public async Task<IActionResult> Login([FromBody] LoginViewModel model)
+        {
+            var result = await _signInManager.PasswordSignInAsync(model.Username, model.Password,
+                model.RememberMe, lockoutOnFailure: false);
+
+            if (result.Succeeded)
+            {
+                return Ok();
+            }
+            if (result.IsLockedOut)
+            {
+                return Forbid();
+            }
+
+            return BadRequest();
+        }
+    }
+}
